+++ pkg/services/control/server/evacuate.go
@@ -68,7 +68,7 @@ func (s *Server) replicate(addr oid.Address, obj *objectSDK.Object) error {
 
 	ns, err := nm.ContainerNodes(c.PlacementPolicy(), cid)
 	if err != nil {
-		return fmt.Errorf("can't build a list of container nodes")
+		return fmt.Errorf("can't build a list of container nodes: %w", err)
 	}
 
 	nodes := placement.FlattenNodes(ns)